# Faster, lock-free pool