--- a/.travis.yml
+++ b/.travis.yml
@@ -4,7 +4,6 @@ services:
 - redis-server
 
 go:
-  - 1.2
   - 1.3
   - 1.4
   - tip