Before:
```
$ go test ./... -run=NONE -bench=Pool -benchmem -cpu=1,2,4,8 -benchtime=2s
BenchmarkPool   10000000         293 ns/op         0 B/op        0 allocs/op
BenchmarkPool-2  3000000        1127 ns/op         0 B/op        0 allocs/op
BenchmarkPool-4  2000000        1187 ns/op         0 B/op        0 allocs/op
BenchmarkPool-8  2000000        1459 ns/op         0 B/op        0 allocs/op
```

After:
```
$ go test ./... -run=NONE -bench=Pool -benchmem -cpu=1,2,4,8 -benchtime=2s
BenchmarkPool   10000000         330 ns/op         0 B/op        0 allocs/op
BenchmarkPool-2 10000000         288 ns/op         0 B/op        0 allocs/op
BenchmarkPool-4 10000000         401 ns/op         0 B/op        0 allocs/op
BenchmarkPool-8 10000000         403 ns/op         0 B/op        0 allocs/op
```