@vmihailenco 2015-02-02
Thanks! Finally package got some docs :)